dataset/cwe_434_unrestricted_file_upload_testcases/testcase2_mime_type_spoofing/requirements.txt
flask==2.0.1
werkzeug==2.0.1
python-magic==0.4.27 